requirements.txt

Django==3.0.11
django-adminlte2-templates==1.4.0
django-allauth==0.44.0
django-celery==3.3.1
django-crispy-forms==1.10.0
django-jalali==4.1.2
django-jalali-date==0.3.1
jdatetime==3.6.2
packaging==20.8
Pillow==8.0.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
pyparsing==2.4.7
Pyphen==0.10.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
txaio==20.12.1
urllib3==1.26.2
WeasyPrint==51
webencodings==0.5.1
wrapt==1.12.1

zope.interface==5.2.0
